msm: mdss: avoid clocks on if panel is not on for command mode
There are cases where vsync enable may happen before panel has been
turned on. In this case MDP clocks and specially DSI clocks shouldn't
get turned on if the panel is not ready yet. Add a check to ensure
clocks are not turned on and wait for panel to be turned on later on,
vsync events will be enabled then.

Change-Id: Id43adfadc9069a5fb56517b326f197a1c9de152b
Signed-off-by: sudosurootdev <sudosurootdev@gmail.com>